menu: hide install options on installed machine
If a machine has been installed with dd, like VMs,
hide the install options.
The fix should work also with machine with extra external storage.

The install option can still appear when connecting an extra storage
which is not already mounted. This should be a temporary
situtation and culd be ignored.