ns-migration: set mwan last_resort to 'default'
The default value for 'last_resort' is 'unreachable'.
If all policies have 'last_resort' set to 'unreachable' and all wans
go down, there is no way to restore the wan connection.

The migration now sets `last_restort` to `default`: if a wan is down,
the traffic tries to use the default gateway.