ns-openvpn: fix check-status for migrated users
Migrated users from an account provider have differences compared to regular users.
While the username is simple, such as "goofy," the Common Name (CN) in the certificate
includes the domain, like "goody@test.local."

This commit modifies the check-status script to account for this condition when
testing the enabled/disabled status for migrated users.