ns-report: dpireport, wait for netifyd socket
On slow machines, netifyd could take a while to start.
Add a wait loop inside dpireport reader thread to wait for the socket:
timeout is 60 seconds.

A waitloop inside the init.d script is not possible
because procd do not start the execution of next scripts until
the procd_open_instance is called.
A wait loop added before starting the application blocks the entire
init process.