ns-dpi: restore signatures
Signature files are not part of backup and get
lost on upgrade and restore.

Involved files:
- /etc/netify.d/netify-categories.json
- /etc/netify.d/netify-apps.conf

These files are updated every day.
If such files are added to the backup, the backup will
be recreated and sent to the remote server
every night.

To avoid such scenario, files are not saved
but downloaded again at every boot that occurs
on upgrade and restore.