ns-threat_shield: force banip restart
When changing ban_ifv4, ban_ifv6, ban_dev and ban_trigger options,
after the banip database commit, the nft rules are not updated.

This change is an hack to force the restart of banip.
A simple "/etc/init.d/banip restart" is not enough, because
sometimes the service does only a reload which does not
recreate nft rules