ns-migration: import, log to persistent storage
If the migration fails, there is no log available.
Make sure to log all executed commands along with stdout and
stder to /root/migration.log which is inside a persistent directory.

Also, the presence of /root/migration.log file is a flag that
the system has executed the migration procedure at least once.