fix(snort3): download rules on startup
When a machine is rebooted or updated using an image, Snort rules are
not present because stored in RAM.

During the service start, if rules are not present,
the init script now tries to load the rules multiple times.
Maximum wait time is 1 minute.

Snort will not fail the start: a nightly cron job
can update the rules and restart the service.