fix(ns.ha): dhcp, check for dns option router
As default, dnsmasq sets the DNS server as the address of the machine
running dnsmasq itself: this will not work correctly for clients
provisioned with DHCP.

Make sure the `dns` option is set if DHCP server is enabled.
The value of the option can be any IP address.

Also do not check if IPv6 DHCP server is enabled: DHCPv6 is currently
not supported.