feat(ns-ha): allow custom lan and wan interfaces
Previous implementation was assuming the name of lan and wan device
was fixed: this was preventing adding the HA feature on existing
firewalls.

Now the config script and the API allow a custom interface name for lan
and wan.
Also please note that wan is not checked on backup node, because
the switchover will work as long as the underlaying device
of the wan (eg. eth1) is the same on both nodes.