fix(ha): export/import DHCP leases
Even if /tmp/dhcp.leases has been added to the sync list
using add_sync_file function, it's not synched.
It seems that the rsync script only sync the following path:
- anything listed inside 'sysupgrade -l'
- anything listes inside the sync_list option

If a new file is added inside the sync_list option, the file
is copied over to the secondary node inside /usr/share/keepalived/rsync
directory, but it's not restored inside the original location
in case of master switch.

To workaround this bug, the DHCP lease file is now
handled from the ns-ha-export and ns-ha-import scripts.

Also add a fix to the export/import scripts to avoid failure
if files do not exists.