fix(ns-ha): do not set disabled flag
Due to a race conditions, interfaces and routes
could be exported to disabled by the primary node:
as a side such interfaces and routes could not be re-actived
during the switch phase.

The race condition was due to the execution of ns-ha-export, before
the 'disabled' option was reverted to '0' on the primary node.

This commit introduced the following changes:

- never set the 'disabled' option
- all wan, wireguard and ipsec interfaces are explictly controlled
  with ifdown/ifup
- reload_config is now called only be ns-ha-enable script
- routes will be active also on the secondary node